# perf: Only call as_polars_selector.character once inside `parse_into_selector`